% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_from_server}
\alias{delete_from_server}
\title{Delete packages}
\usage{
delete_from_server(monorepo_url, cranlike_url)
}
\arguments{
\item{monorepo_url}{git url of the monorepo}

\item{cranlike_url}{url to the cranlike server packages api}
}
\description{
Delete packages from the server.
}
