% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_from_server}
\alias{delete_from_server}
\title{Delete packages}
\usage{
delete_from_server(universe)
}
\arguments{
\item{universe}{name of the universe}
}
\description{
Delete packages from the server.
}
